feat(runtimed-py): add outputs to Cell class for cross-session visibility
The Cell class now includes an `outputs` field that is populated when
calling get_cell() or get_cells(). This enables agents to see historical
execution outputs from cells executed by other clients.

- Add `outputs: Vec<Output>` field to Cell struct with pyo3 getter
- Create shared output_resolver module for blob hash resolution
- Update get_cell/get_cells in Session and AsyncSession to resolve outputs
- Remove duplicate output parsing code from session modules

Resolves: outputs not visible in Cell class when retrieved from Automerge
